// dashboard/src/utils/hasura-api/generated/schemas/deleteEventTriggerBulkOperation.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { DeleteEventTriggerBulkOperationType } from './deleteEventTriggerBulkOperationType';
import type { DeleteEventTriggerStep } from './deleteEventTriggerStep';

export interface DeleteEventTriggerBulkOperation {
  /** Execute multiple operations in a single query */
  type: DeleteEventTriggerBulkOperationType;
  /** Source of the operation */
  source: string;
  /** Resource version of the operation */
  resource_version?: number;
  args: DeleteEventTriggerStep[];
}